+++ river_flows/orm/site_condition.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import DateTime, Integer, String
+from sqlalchemy import DateTime, Float, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from river_flows.orm.base import Base
@@ -12,5 +12,5 @@ class SiteCondition(Base, TimestampMixin):
     site_id: Mapped[str] = mapped_column(String, primary_key=True)
     site_name: Mapped[str] = mapped_column(String)
     timestamp: Mapped[datetime] = mapped_column(DateTime(timezone=True), primary_key=True)
-    value: Mapped[int] = mapped_column(Integer)
+    value: Mapped[int] = mapped_column(Float)
     unit: Mapped[str] = mapped_column(String)
